Add thirdMax function in 48.js
This commit adds a new function, thirdMax, to 48.js. The function takes an array of numbers as an input and returns the third highest number. If there is no third max, the function returns the max instead. This function was created as a response to a problem found on LeetCode and provides a utility for finding the 3rd highest number in a list of numbers.