Add solution for finding disappeared numbers in array
Added a new JavaScript file '49.js', implementing the function 'findDisappearedNumbers'. This function aims to solve a LeetCode problem where it finds all the elements of [1, n] inclusive that do not appear in a given array. The function uses the absolute value of each element in the array as an index in a loop and inverts the element at that index. The result array is then populated with indices of positive values in the array, incremented by 1.