Add 'last' method to Array prototype
Added a new method 'last' to the Array prototype for enhancing the library's capabilities. This method returns the last element of an array if it exists, otherwise, it gives -1. This could be useful in scenarios where we need to check the last element frequently. The method was tested with an array [1,2,3] and worked as expected.